refactoring: factor out workspace creation, reuse for organize imports
